requirements.txt
Flask==2.0.2
Flask-PyMongo==2.3.0
pysnmp==4.4.12
pytest==6.2.5
